Change int to long to help with hibernate identity

+++ src/main/java/fr/miage/agents/api/model/Categorie.java
@@ -7,7 +7,7 @@ import java.io.Serializable;
  */
 public class Categorie implements Serializable {
 
-    public int idCategorie;
+    public long idCategorie;
 
     public String nomCategorie;
 }
